[BE-135] Session Serializer 타입 에러 수정

+++ src/main/java/com/recordit/server/util/SessionUtil.java
@@ -23,13 +23,13 @@ public class SessionUtil {
 	}
 
 	public Long findUserIdBySession() {
-		Long userId = (Long)httpSession.getAttribute(PREFIX_USER_ID);
+		Integer userId = (Integer)httpSession.getAttribute(PREFIX_USER_ID);
 		if (userId == null) {
 			log.info("세션에 사용자 정보가 저장되어 있지 않습니다");
 			invalidateSession();
 			throw new NotFoundUserInfoInSessionException("세션에 사용자 정보가 저장되어 있지 않습니다");
 		}
-		return userId;
+		return userId.longValue();
 	}
 
 	public void isCorrectSession() {
